requirements.txt
africastalking==1.2.8
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
awscli==1.34.22
awsebcli==3.21.0
botocore==1.35.22
CacheControl==0.14.0
cachetools==5.5.0
cement==2.10.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
django-auth-adfs==1.14.0
django-oauth-toolkit==3.0.1
django-oidc-provider==0.8.2
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.16
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
ecdsa==0.19.0
firebase-admin==6.5.0
future==1.0.0
gcloud==0.18.3
git-filter-repo==2.45.0
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jose==1.0.0
josepy==1.14.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
jwks==1.0.0
jwt==1.3.1
mozilla-django-oidc==4.0.1
msgpack==1.1.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pathspec==0.10.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyjwkest==1.4.2
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
Pyrebase4==4.8.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-jwt==4.1.0  # Consider checking for updates for compatibility
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.2
schema==0.7.7
semantic-version==2.10.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
